snippets/requirements.txt
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
djangorestframework==3.10.3
gunicorn==20.0.4
psycopg2==2.8.4
Pygments==2.5.2
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
